[BAU] combine two validation utils test modules
